+++ src/__tests__/useForm/resolver.test.tsx
@@ -1,5 +1,11 @@
 import * as React from 'react';
-import { fireEvent, render, screen } from '@testing-library/react';
+import {
+  act,
+  fireEvent,
+  render,
+  screen,
+  waitFor,
+} from '@testing-library/react';
 
 import { useForm } from '../../useForm';
 
@@ -46,5 +52,112 @@ describe('resolver', () => {
     fireEvent.click(screen.getByRole('button'));
 
     screen.findByText("{test:'test'}");
+  });
+
+  it('should support resolver schema switching', async () => {
+    type FormValues = {
+      test: string;
+    };
+
+    const fakeResolver = (schema: boolean) => async () => {
+      return schema
+        ? {
+            values: { test: 'ok' },
+            errors: {},
+          }
+        : {
+            values: {},
+            errors: {
+              test: {
+                type: 'test',
+                value: { message: 'wrong', type: 'test' },
+              },
+            },
+          };
+    };
+
+    const App = () => {
+      const [schema, setSchema] = React.useState(false);
+      const [submit, setSubmit] = React.useState(false);
+      const {
+        register,
+        handleSubmit,
+        formState: { errors },
+      } = useForm<FormValues>({
+        resolver: fakeResolver(schema),
+      });
+
+      return (
+        <form
+          onSubmit={handleSubmit(() => {
+            setSubmit(true);
+          })}
+        >
+          <input {...register('test')} />
+          {errors.test && <p>Error</p>}
+          {submit && <p>Submitted</p>}
+          <button onClick={() => setSchema(!schema)}>Toggle</button>
+          <button>Submit</button>
+        </form>
+      );
+    };
+
+    render(<App />);
+
+    await act(async () => {
+      fireEvent.click(screen.getByRole('button', { name: 'Submit' }));
+    });
+
+    await waitFor(async () => {
+      screen.getByText('Error');
+    });
+
+    await act(async () => {
+      fireEvent.click(screen.getByRole('button', { name: 'Toggle' }));
+    });
+
+    await act(async () => {
+      fireEvent.click(screen.getByRole('button', { name: 'Submit' }));
+    });
+
+    await waitFor(async () => {
+      screen.getByText('Submitted');
+    });
+  });
+
+  it('should be called with the shouldUseNativeValidation option to true', async () => {
+    const test = jest.fn();
+    const resolver = (a: any, b: any, c: any) => {
+      test(a, b, c);
+      return {
+        errors: {},
+        values: {},
+      };
+    };
+
+    const App = () => {
+      const { register, handleSubmit } = useForm({
+        resolver: async (data, context, options) =>
+          resolver(data, context, options),
+        shouldUseNativeValidation: true,
+      });
+
+      return (
+        <form onSubmit={handleSubmit(() => {})}>
+          <input {...register('test')} />
+          <button>Submit</button>
+        </form>
+      );
+    };
+
+    render(<App />);
+
+    await act(async () => {
+      fireEvent.click(screen.getByRole('button'));
+    });
+
+    expect(test.mock.calls[0][2]).toEqual(
+      expect.objectContaining({ shouldUseNativeValidation: true }),
+    );
   });
 });