+++ src/useController.ts
@@ -40,14 +40,13 @@ export function useController<
     inFieldArrayActionRef,
   } = control || methods.control;
 
-  const isFieldArray = isNameInFieldArray(namesRef.current.array, name);
   const field = get(fieldsRef.current, name);
   const [value, setInputStateValue] = React.useState(
-    isFieldArray || !field || !field._f
-      ? isFieldArray || isUndefined(get(defaultValuesRef.current, name))
-        ? defaultValue
-        : get(defaultValuesRef.current, name)
-      : field._f.value,
+    field && field._f && !isUndefined(field._f.value)
+      ? field._f.value
+      : isUndefined(get(defaultValuesRef.current, name))
+      ? defaultValue
+      : get(defaultValuesRef.current, name),
   );
   const { onChange, onBlur, ref } = register(name, {
     ...rules,
@@ -65,18 +64,28 @@ export function useController<
         setInputStateValue(get(data.values, name)),
     });
 
+    const filed = get(fieldsRef.current, name);
+
+    if (field && field._f) {
+      filed._f._c = true;
+    }
+
     return () => {
       controllerSubscription.unsubscribe();
       const shouldUnmountField = shouldUnmount || shouldUnregister;
 
       if (
-        isFieldArray
+        isNameInFieldArray(namesRef.current.array, name)
           ? shouldUnmountField && !inFieldArrayActionRef.current
           : shouldUnmountField
       ) {
         unregister(name);
-      } else if (field && field._f) {
-        field._f.mount = false;
+      } else {
+        const field = get(fieldsRef.current, name);
+
+        if (field && field._f) {
+          field._f.mount = false;
+        }
       }
     };
   }, [name]);